Phase 3.1: Extract formatting functions to searcher_utils.py
- Create doc_search/searcher_utils.py module with:
  - highlight_terms(): Mark up search terms with text markers
  - highlight_terms_ansi(): Mark up search terms with ANSI colors
  - check_phrase_match(): Check if phrase appears in text
  - find_best_snippet(): Extract most relevant snippet from text
  - format_results(): Format search results for display

- Update searcher.py to import from searcher_utils
- Remove duplicate function definitions from searcher.py
- Keep find_phrase_positions() in searcher.py (query parsing)

All 427 tests pass.
